# services/pattern-manager/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Pattern Manager API",
    "description" : "Pattern Store read API + human-approval lifecycle for discovered correlation patterns (RCA, structural validation, codebook reconciliation, session-window, XAI).",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/patterns/{patternId}/deprecate" : {
      "post" : {
        "tags" : [ "pattern-controller" ],
        "summary" : "Deprecate a draft or approved pattern",
        "operationId" : "deprecate",
        "parameters" : [ {
          "name" : "patternId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeprecateIntent"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PatternView"
                }
              }
            }
          }
        }
      }
    },
    "/patterns/{patternId}/approve" : {
      "post" : {
        "tags" : [ "pattern-controller" ],
        "summary" : "Approve or reject a draft pattern",
        "operationId" : "approve",
        "parameters" : [ {
          "name" : "patternId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ApprovalIntent"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PatternView"
                }
              }
            }
          }
        }
      }
    },
    "/patterns/{patternId}" : {
      "get" : {
        "tags" : [ "pattern-controller" ],
        "summary" : "Get a single pattern by id (full PatternView incl. XAI + sessionWindow)",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "patternId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PatternView"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "pattern-controller" ],
        "summary" : "Edit a draft pattern (per-position optional markers; frozen PatternEdit body)",
        "operationId" : "edit",
        "parameters" : [ {
          "name" : "patternId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatternEdit"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PatternView"
                }
              }
            }
          }
        }
      }
    },
    "/patterns" : {
      "get" : {
        "tags" : [ "pattern-controller" ],
        "summary" : "List patterns (PatternPage envelope), filterable by lifecycle",
        "operationId" : "list",
        "parameters" : [ {
          "name" : "lifecycle",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PatternPage"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "DeprecateIntent" : {
        "properties" : {
          "reviewer" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          }
        },
        "required" : [ "reviewer" ]
      },
      "JsonNode" : { },
      "PatternView" : {
        "properties" : {
          "patternId" : {
            "type" : "string"
          },
          "trailId" : {
            "type" : "string"
          },
          "sequence" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SequenceElementView"
            }
          },
          "rootCauseAlarmType" : {
            "type" : "string"
          },
          "support" : {
            "type" : "number",
            "format" : "double"
          },
          "confidence" : {
            "type" : "number",
            "format" : "double"
          },
          "lift" : {
            "type" : "number",
            "format" : "double"
          },
          "timing" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "sessionWindow" : {
            "$ref" : "#/components/schemas/SessionWindowView"
          },
          "codebookMatchId" : {
            "type" : "string"
          },
          "reconcileStatus" : {
            "type" : "string"
          },
          "structurallyValidated" : {
            "type" : "boolean"
          },
          "structuralValidationReason" : {
            "type" : "string"
          },
          "instanceCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "supportingInstances" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SupportingInstanceView"
            }
          },
          "lifecycle" : {
            "type" : "string"
          },
          "domain" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "SequenceElementView" : {
        "properties" : {
          "alarmType" : {
            "type" : "string"
          },
          "optional" : {
            "type" : "boolean"
          }
        }
      },
      "SessionWindowView" : {
        "properties" : {
          "windowMs" : {
            "type" : "integer",
            "format" : "int64"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
      "SupportingInstanceView" : {
        "properties" : {
          "sourceWindowId" : {
            "type" : "string"
          },
          "snapshotId" : {
            "type" : "string"
          },
          "occurrence" : {
            "$ref" : "#/components/schemas/JsonNode"
          }
        }
      },
      "ApprovalIntent" : {
        "properties" : {
          "decision" : {
            "type" : "string",
            "pattern" : "approve|reject"
          },
          "reviewer" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          }
        },
        "required" : [ "decision", "reviewer" ]
      },
      "PatternEdit" : {
        "properties" : {
          "sequenceFlags" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SequenceFlag"
            }
          },
          "reviewer" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          }
        },
        "required" : [ "reviewer", "sequenceFlags" ]
      },
      "SequenceFlag" : {
        "properties" : {
          "index" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "optional" : {
            "type" : "boolean"
          }
        },
        "required" : [ "index", "optional" ]
      },
      "PatternPage" : {
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PatternView"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          },
          "offset" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  }
}